Break out Constants and RouteControllerDelegate
